Fixed two spinner Content/Control view issues.
- Issue #272 ControlViewSpinnerStructure should allow ListItem
- Issue #268 ContentViewSpinnerStructure should allow ListItem

In both cases, the code now allows for 0 or more ListItem children of the spinner control.

Note: Keen observers of the documentation may notice that list items are permitted only when the Selection control pattern is present, and such a check is not done in this code. This is a shortcoming of several of the content/control view rules. However, since no rules currently check for such patterns, it is not the intent to remedy that as part of this change.

Another such failing is that the children are not actually checked to make sure they have the IsContentElement or IsControlElement properties set to true. This failing is common throughout these rules. And again, there is no intent to remedy that as part of this change.