Remove CCAControlTypeFilters from Axe.Windows.Rules
The class CCAControlTypeFilters was unnecessary because the integer values of the UIA control types could more directly be enumerated by the ControlType class. In addition, CCAControlTypeFilters was removed because it was made part of the Rules assembly, whose responsibility is only to run automated rules -- the Rules assembly should never know anything about the color contrast analyzer, or report any type information other than that provided by the rules.