Hide patterns treeview when no patterns are present (#505)
Currently, when no patterns are present, we display, "No available pattern..." in place of the patterns tree. For keyboard users, however, focus still goes to the empty tree rather than the text. This PR updates the keyboard nav behavior so that focus goes to the text instead of the empty tree. This is accomplished by a. hiding the tree and b. making the text focusable.