Keep textbox placeholder text on keyboard focus (#508)
Currently, we have some textboxes with placeholder text. This text disappears when the textbox receives keyboard focus. This PR updates that behavior such that the placeholder text instead disappears only when the textbox is not empty, regardless of keyboard focus. A new PlaceholderTextBox class extending TextBox is introduced to accomplish this.