Set dummy HelpText value on some buttons (#519)
According to the docs, AutomationProperties.HelpText is generally the same text that is provided in the tooltip for the control. WPF takes this to heart and automatically sets a button's helptext to its tooltip value. Frustratingly, if the tooltip is set as a text-containing control such as a Run, the helptext is set to be that control's class name. For ai-win's icon buttons, the tooltip's content is roughly the same as the AutomationProperties.Name's content. To avoid repetitive screen reader output, we don't want to set the helptext with the tooltip at all (much less with the tooltip's class name). This PR sets dummy (blank space) values to the helptext in the situations where we want to stop the default tooltip-as-helptext behavior.