Fix signed build (#536)
Recent changes to our UI tests (#528) broke our signed build. Specifically, the change to make Win32's Internals visible to UITests broke because we don't sign the UITests project (because some of its dependencies aren't signed). This PR addresses the problem by:
a. No longer trying to sign the UI test job
b. Introducing (and using) a new SignedRelease configuration, which is the same as Release except that it excludes UITests