change light gray color (#533)
This PR changes TextBrushGray from #737373 to #6E6E6E. This changes the contrast ratio > 5:1. From Chelsea:

If we brought the secondary text in Accessibility Insights for Windows to #6E6E6E, that would bring the ratio up to 5.09:1. If that's not a high enough bump, let me know and I'll tweak the color some more