Change to HelpText what was mistakenly set to AutomationName (#539)
This PR picks up some stuff that was missed in PR #525. Description copied from there:
The help text for the control was accidentally set as the name, making the text "sample small bold text". The tester was concerned the screen reader user would not hear the text as it appears on screen. Now the name is "the quick brown fox" and the help text is "sample small bold text".
A resource name was also updated while in the neighborhood.