Replace Properties DataGrid with ListView (#541)
This PR replaces the Properties DataGrid with a ListView/GridView combination. This is done to improve the screen reader experience of the grid. It should be visually nearly identical. The key difference is that a user can now only navigate through entire rows, not cells. Extensive customization was added to match the existing visual appearance, column sizing, and keyboard behavior.