Update TextBox event handling (#553)
Picking up two minor changes missed in PR #552. Two event handler methods were switched from a window to a textbox, but their names still referenced the window. This PR renames them accordingly. It also prevents the Cut and Delete commands from being executed on the textbox, since it is not meant to be directly edited.