Don't show welcome screen when opening loaded file (#616)
To address #608- [BUG] Opening file by file association is broken if welcome dialog is shown, this PR stops showing the startup/welcome screen when loading a file directly on launch. This makes more sense to me than our old behavior (which I would describe as unintentional), which was that the welcome screen would show over the loaded data. If others disagree, we can revert to that behavior instead.