Gate signing only with build variables (#622)
PRs #536, #620, and #621 failed to get us digitally and strong name signed. Hopefully this PR will. Here, we stop gating signing based on the Configuration and leave it entirely up to the build variables. We also tweak the variables set in signedbuild.yml to make signing opt-in instead of opt-out. A signed build ran and produced a digitally and strong name signed artifact.