Let a background in CCA mode revert to default with high contrast on (#680)
A recent issue (#675) pointed out that a certain control in the color contrast mode had poor color contrast while using high contrast themes. This PR addresses this issue by removing the specific color we set for that control's background in HC modes. This allows the background to revert to the control's default, which offers proper contrast.