Set minwidth in mode control left columns (#852)
Currently, if a user reduces a left column width (such as the hierarchy tree in live mode) to 0, they can no longer use the mouse to resize that column. This PR addresses that concern by setting the min column width of applicable columns to 2. This ensures the user can always resize the column with their mouse.