Stop left column from extending outside of window (#899)
As promised in #852, this PR introduces a max column width to our main UI's left column (which usually holds the UIA tree). This is analogous to the min column width requested in #814. To accomplish this, ColumnMaxWidthSpacingConverter is introduced and used to limit the column's max width to slightly smaller than its parent control. The LeftPaneColumn style is introduced and used to apply this converter to the left columns in the Live, Event, and Test (Snapshot) mode controls.