Fix: Use correct brush for TreeView Expand/Collapse icons when selected (#963)
In HC Light and HC Dark, the Expand/Collapse icon in our TreeViews fails contrast when the item is selected. This PR addresses #860 (which calls out the UIA TreeView) as well as our other TreeViews, such as the Patterns tree. It sets the color of the expander to match the color of the other text and icons on selected items.