fix: Improve contrast of test tab header focus indicator with inner border (#969)
Addresses Focus indicator on 'selected' test in Fast Pass has insufficient contrast against its background #868. Currently, the focus indicator on a test tab header is the normal black or white outline, but the color of the selected tab can be close to black or white. We had a similar problem with our left nav bar buttons, so I used the same solution here. Now, when a test tab has focus, it gets an inner border with the opposite (black or white) color of the current focus indicator. This means that there will always be a black and white border to indicate focus, which will pass contrast requirements. Also, this active rectangle on the left side of the selected tab now has a better color for high contrast modes.