fix: Replace tab stops help icon with text (#1052)
Details
Currently, tab stops has a help (i) icon that is a hyperlink. It does not show a tooltip on keyboard focus. This PR replaces that icon with text to avoid the need for a tooltip.

Motivation
Addresses an accessibility issue

Context
I tried to adapt KeyboardToolTipButtonBehavior to function on hyperlinks. Unfortunately, Hyperlink is not a FrameworkElement and thus cannot be set as a ToolTip's PlacementTarget and does not prove screen location information. This makes it very difficult to position a keyboard focus triggered ToolTip on the screen for a Hyperlink. As such, this PR's approach was selected instead.