fix: Set automation name on CustomGridViewColumn headers (#1072)
Details
Currently, our properties table column headers are visible and directly accessible via assistive technology as TextBlock elements. Unfortunately, the header elements exposed by the table patterns are not those TextBlock elements and do not have accessible names set. This PR addresses the issue by setting AutomationProperties.Name on those header elements to the same value has the text displayed.

Motivation
Fixes an accessibility issue.

Context
I tried a few different ways to do this without success, including setting the Header property on the table directly, setting LabeledBy on the header to the visible text element, and changing the TextBlock we use to render the text to be a Label. I bet there are other ways to achieve the desired behavior, but this was simple and successful so I stopped searching.