fix: Convert the update dialog to be a ContainedDialog (#1201)
#### Details
Currently, our update dialog is a traditional modal dialog in a separate window. This results in some undesirable side effects while the dialog is open, such as the inability to reposition or close the application. To improve this behavior, this PR switches the update dialog to be a `ContainedDialog`, following the same pattern as our telemetry and channel switching dialogs. With one exception, the updated `UpdateContainedDialog` provides exactly the same functionality to the user as the existing `UpdateDialog`. The exception: for required updates, there is now a Close button in place of the Later button for option updates.

##### Motivation
Addresses issue #1186 (at least partially) 

##### Context
- To support this change, I updated `DialogContainerModeControl` to disable all background `ContainedDialog`. This way, in the (rare) scenario where the user has the Update, Telemetry, and StartUp dialogs all open at once, only the topmost contained dialog will show up in the tab order.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Run through of all [test scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md) completed?
- [x] Does this address an existing issue? If yes, Issue# - #1186
- [x] Includes UI changes?
  - [x] Run the production version of Accessibility Insights for Windows against a version with changes.
  - [x] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked off. All of these checks must pass before a merge.