fix: prevent crash in Highlighter beaker rendering due to FontFamily use-after-dispose (#1291)
#### Details

This PR fixes issue #1290, where one user reported an issue where Application Insights would crash while trying to render the "beaker" icon as part of its highlighter overlay.

The root cause of the crash is that the code path in question was rendering the icon using a `FontFamily` pulled out of a `PrivateFontCollection` which we had already `Dispose()`ed. The `Families` getter in the `PrivateFontCollection` doesn't document the lifetime requirements of the `FontFamily` objects it returns and its implementation is misleading in that it *looks* like it's cloning a copy of the underlying native font objects with a separate lifetime, but the native clone still maintains some links to data backed by the font collection, so it isn't safe to use it this way in practice.

##### Motivation

Addresses issue #1290

##### Context

I considered a more contained version of this fix which just expanded the scope of the `using` block that we were previously using to control the lifetime of the `PrivateFontCollection` to the end of the function it's used in. I went with the approach here instead because it solves the same problem but also has the benefit of moving the slightly expensive "load and parse the font" operation out of the middle of a rendering/window message handling path and into a "only once during startup" path instead, which results in a marginally more fluid user experience.

I also included a change here that changes an existing exception handling path to report any exceptions that occur while loading the font, per discussion last month with @DaveTryon. This didn't turn out to be relevant to this bug, but I kept it since it seemed like a good change to make.

We don't have a local repro of the issue to test against, but I did send an earlier version of this PR to the originally reporting user to verify, and they confirmed that it fixed the issue. Once I've addressed any initial review comments, I'll send that user a copy of the final candidate fix to double check that the fix still works as expected.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [ ] Run through of all [test scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md) completed?
- [x] Does this address an existing issue? yes, #1290
- [no] Includes UI changes?
  - [n/a] Run the production version of Accessibility Insights for Windows against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked off. All of these checks must pass before a merge.