fix: Bold MenuItem text when selected (#1321)
#### Details
Currently, some of our MenuItems are bolded and highlighted when selected, while others are just highlighted. Because the default highlight behavior has insufficient color contrast, this PR ensures all MenuItems are bolded when selected. To do this, 
`menuPlainWhite` is applied to as many menus as possible. In cases where it has issues, a new `menuDefault` style has been created to provide the bolding behavior. 

Screen capture:
![screen capture showing a variety of menus with the new bolding on selection behavior](https://user-images.githubusercontent.com/4615491/164778387-d94879fa-807d-4095-a134-2144258738d7.gif)

##### Motivation
Addresses an accessibility issue

##### Context
Ideally, we would fully standardize to use `menuPlainWhite`. Unfortunately, it has issues with multi-leveled menus which were too complicated to address as a part of this PR.

To reduce the amount of size changing on selection, a minor `LayoutTransform` is applied alongside the bolding.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Run through of all [test scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md) completed?
- [x] Does this address an existing issue? If yes, Issue# - ADO 1928590
- [x] Includes UI changes?
  - [x] Run the production version of Accessibility Insights for Windows against a version with changes.
  - [x] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked off. All of these checks must pass before a merge.