chore: Remove DisableDarkMode config setting (#1419)
#### Details

Remove support for the `DisableDarkMode` config setting. This setting has no UI and required manual editing of the config file.

##### Motivation

When dark mode was introduced in #719, it was gated behind a non-UI config setting named `EnableDarkMode` so that people had to explicitly _opt in_ to use dark mode. When dark mode was enabled by default in #764, we provided a non-UI config setting named `DisableDarkMode` so that people had a way to _opt out_ from dark mode in AIWin. Both of these settings were included because we were concerned about dark mode bugs that could potentially block usability. That was 2 years ago and dark mode has matured a lot since then, so this config setting can now be removed.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [ ] Run through of all [test scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md) completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [ ] Includes UI changes?
  - [ ] Run the production version of Accessibility Insights for Windows against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked off. All of these checks must pass before a merge.