chore: Target .NET Framework 4.8 (#1465)
#### Details

We need .NET Framework 4.8 to address some accessibility issues, because
the required properties are not exposed in .NET 4.7.2. Telemetry
indicates that 98.1% of our current users are running on systems where
.NET Framework 4.8 is already installed and will not be impacted by this
change. The remaining 1.9% will experience a failed upgrade until they
install a newer version of .NET Framework 4.8. They can learn of this
requirement in any of the following ways:
1. Read the release notes (accessible directly from the upgrade dialog)
2. Uninstall and reinstall AIWin (the installer will tell the user to
install the new version)
3. Search for a GitHub issue that we will create after this PR merges

Note: The primary change is captured in the change to
`build\NetFrameworkAll.targets`, then the other changes are the
downstream effects or doc changes. The notable exception is that
`CustomActions.Package.csproj` uses the old-format project file (WiX
doesn't play nicely with the new format), so it needed to be explicitly
updated as well. This was done via the IDE.

##### Motivation

Enable `AutomationProperties.SizeOfSet` and
`AutomationProperties.PositionInSet` properties to address an
accessibility issue.

##### Screenshots

The only differences are when running the installer, which informs the
user which runtime is needed:

Old | New
--- | ---

![image](https://user-images.githubusercontent.com/45672944/199129698-3bb3bcee-2df6-4686-8a3c-8b9c219ba5ca.png)
|
![image](https://user-images.githubusercontent.com/45672944/199130835-b266c16f-4ddb-4756-8c59-7694c52f68ee.png)

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
Using the new properties will be in a separate PR to keep things clean.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
We considered moving all the way to .NET Framework 4.8.1 (released in
August 2022), but telemetry showed that this would impact about 60% of
our users. Since .NET Framework 4.8 exposes the properties that we need,
we opted for the less intrusive option.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [x] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.