fix: Prevent horizontal shifting in properties grid (#1475)
#### Details

By default when a `ScrollViewer` is wrapping a `Grid`, WPF scrolls to
focus on the element being selected in the grid. This is the behavior we
want for vertical scrolling, but this causes some jumpy behavior in the
horizontal scrollbar, as described in #1150. The horizonal jumps would
make more sense if the user were navigating between cells horizontally,
but since we treat the properties grid as a list and only allow
navigation between rows (as opposed to between cells), the default
behavior just looks odd. As a result, this PR overrides WPF default
behavior by keeping vertical navigation the same but removing the
horizontal component.

##### Motivation

Addresses issue
https://github.com/microsoft/accessibility-insights-windows/issues/1150

##### Context

Behavior after this change:
[Description: Gif of AIWin in live inspect mode. User navigates arounds
property grid with keyboard. Navigation is the same as before but
horizontal jumps have been removed.]

![Accessibility Insights for Windows - Inspect - Live 2022-11-09
09-33-19](https://user-images.githubusercontent.com/16010855/200900956-87ccd20d-a93e-48a4-a276-5530c7d554f3.gif)

#### Pull request checklist

- [x] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [x] Does this address an existing issue? If yes,
https://github.com/microsoft/accessibility-insights-windows/issues/1150
- [x] Includes UI changes?
- [x] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [x] Attach any screenshots / GIF's that are applicable.