chore: Fix gap in WxsValidationTests (#1483)
#### Details

This fixes some issues with the MSI checker--I discovered these when I
added a file to the VersionSwitcher but the test still succeeded:
1. Due to a logic error, files were returned _only_ if a set of excluded
files was provided. This empty set was not caught as a problem. These
have both been fixed.
2. We weren't passing the correct component names in 2 places, so we
were returning empty sets of files. This empty set was also not caught
as a problem. These have been fixed.
3. As written, the checker assumed that everything in a component comes
from the same build folder, which we intentionally don't do because of
signing requirements. Now we ignore the path and use just the file name
(with extension) to ensure that the correct set of files is found.

##### Motivation

Make test detect errors (which is why it's there)

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.