chore: Be smarter about recompiling (#1486)
#### Details

The `Win32` project generates a couple of files on the fly, and these
files are then included in all of the non-test projects. This works
nicely for building but has an unfortunate side effect. These files get
generated on _every compile_, which causes the _entire solution_ to
recompile on _every compile_. Since a compile is on the order of 20
seconds, the inner debugging loop is slower than it ought to be,
especially when you're working on unit tests. This PR does the
following:
- Since the attributes are always the same, move them into a static
source file that gets automatically included into every project that we
ship.
- Only generate `A11yInsightsVersionInfo.cs` if it either does not exist
or has changed, and only when building (not cleaning)

Validation scenarios:
- Build with a clean environment (ran `git clean -fdx` before starting
VS). Build succeeds
- Manually edit the generated files and build again. Generated files get
rebuilt
- Rebuild. Generated files do not get rebuilt
- Modify the .csproj file to simulate a different version, then Rebuild.
Generated files get rebuilt

##### Motivation

Speed up inner dev loop

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.