feat(upgrader-security): Use redirectable links to obtain the upgrade manifests (#1490)
#### Details

We currently use hardcoded links to fetch our upgrade manifests. We're
changing these to fwlink values that we can later redirect to the signed
manifests once that work is complete. For the moment, the redirects
point to the same locations as the old hardcoded links. This is easy to
validate by


Tested locally and confirmed that redirects and updates work as
expected.
##### Motivation

Part of the feature to improve security of our upgrader.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
I have a branch with 49 updated files. The hope is to break that set of
changes down into smaller pieces to make it more manageable. You can
confirm that the redirects are the same by copying/pasting the old and
new links into your browser and comparing the resulting JSON that gets
returned.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [ ] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.