feat(upgrader-security) : Generate signed manifest (#1492)
#### Details

Generate the signed manifest file that will eventually replace the
human-curated and unsigned manifest files that we currently use. The
approach is simple:

1. Add a new project that builds `AccessibilityInsights.Manifest.dll`.
This project builds only in release, and only _after_ the MSI has been
fully created
2. Use a script to create a JSON file that contains information about
the generated MSI file
3. Embed this script into the signed manifest assembly. In a future PR,
the client will learn how to extract the signed manifest information and
use it to ensure that we're installing the exact MSI that was built by
the pipeline.

A test iteration of the signed build pipeline can be found
[here](https://dev.azure.com/mseng/1ES/_build/results?buildId=18854875&view=results).
The section dealing with the signed build of the manifest appears
[here](https://dev.azure.com/mseng/1ES/_build/results?buildId=18854875&view=logs&j=7c3ebc89-e7dd-5791-8c02-2ea1f3a27b50&t=9b81c1c0-976a-538f-a1e1-f4d170544876&l=2597).
Full disclosure, I made a couple of tiny changes to the PowerShell
script to make the logging a little clearer. These changes have been
validated locally but will not appear in a signed build until after this
merges.

##### Motivation

Part of feature work to improve our updater security

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
- A future PR will publish the manifest files as a build artifact
- A future PR will include the client-side changes for consuming this
file. The code is written and validated, but I'm trying to keep the PR
sizes as manageable as possible.
- A future PR (after the client code changes are merged) will leverage
the client code in the pipeline to ensure that the client code and the
manifest remain in sync

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [ ] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [ ] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.