feat(upgrader-security): Consume signed manifests, add telemetry and docs (#1493)
#### Details

This adds the client-side code to consume signed manifests (it
temporarily also consumes unsigned manifests), as well as the code to
use the data in the signed manifests to ensure that we only install
non-tampered manifests. I tried to break this PR down into smaller
pieces, but I couldn't get it any smaller than this. A reasonable
overview of the process is added in `ChangingVersions.md`, but here's a
(lengthy) breakdown of the code changes, in rough order of execution:

##### Fetching the manifest
- We add 3 new fields to the `ChannelInfo` class. These fields contain
the metadata about the signed MSI file.
- We define a `StreamMetadata` class to provide more information about
HTTP calls.
- We define an `InterceptingWebClient` class that allows us to capture
the fully resolved Uri that was the actual source of whatever we loaded.
- In `GitHubClient.LoadUriContentsIntoStream`, we use the
`InterceptingWebClient` to build a `StreamMetadata` that describes the
web resource that we fetched.
- In `ChannelIntoUtilities.GetChannelFromStream`, we first try to parse
the stream as a signed manifest (including checking the signature
chain). If that fails, we try to parse the stream as an unsigned
manifest. The code to use the unsigned manifest will be removed after
all release channels have migrated to signed manifests.
- We define an `EnrichedContentInfo` class that derives from
`ContentInfo`, adding the `StreamMetadata` as a member. This gives us a
single object that contains the update information.
- We synthesize `ChannelInfo.MinimumVersion` if we're reading an
unsigned manifest (this data depends on the release channel and is not
included in signed manifest files).
- We add 3 manifest metadata fields to the `IAutoUpdate` interface.
These exist for telemetry, since extensions can't publish telemetry
events directly.
- The `GithubAutoUpdate.AutoUpdate` class populates the metadata
properties based on the data from the `EnrichedChannelnfo` object, and
the manifest metadata fields get bundled with the
`Upgrade_GetUpgradeOption` telemetry action.

##### Running the upgrade
- We define the `ExecutionResult` enum to capture possible outcomes of
running VersionSwitcher.
- We define the `ResultBearingException` class, which is just a regular
`Exception` with an `ExecutionResult` attached.
- We define the `ResultExecutionWrapper` class, which allows us to have
layers of failure capture, and the most granular failure will be
preserved.
- We define an `ExecutionHistory` class that will be used for telemetry
from VersionSwitcher. It also replaces the existing EventLogger and
ExceptionReporter classes that VersionSwitcher used for local debugging.
- The VersionSwitcher's commands have been expanded to accept the MSI
size and SHA, so VersionSwitcher can validate after downloading. For
legacy manifests, these values are unknown. Since VersionSwitcher uses
positional arguments, we need placeholders for the unknown versions. We
use `0` for the unknown size and `none` for the unknown SHA. These new
parameters are passed from `GitHubAutoUpdate.Update` into
`VersionSwitcherWrapper.InstallUpgrade`.
- VersionSwitcher uses any MSI metadata (size and/or SHA) to validate
that the MSI installer is trustworthy before installing it.
- Milestones during the execution of VersionSwitcher get saved in the
`ExecutionHistory` object.
- When VersionSwitcher exits, it serializes the `ExecutionHistory` data
to a known file location.

##### Reporting VersionSwitcher results
- When AIWin starts, it checks the disk for the serialized
`ExecutionHistory`.
- It takes selected data and publishes an
`Upgrade_VersionSwitcherResults` event. This event includes rich
metadata about the MSI file and should allow us to detect any tampering
or corruption of the installer. There's some overlap with the existing
`Upgrade_DoInstallation` telemetry point, so we might want to iterate a
bit on this.
- It renames the file with the serialized `ExecutionHistory`. This
ensures that the telemetry gets updated only once, but keeps the file on
disk for investigateion purposes.

I've confirmed the core scenarios in the
https://github.com/microsoft/accessibility-insights-windows/tree/DaveTryon/add-signed-manifest
branch, but the real test will come when we get it into the Canary
branch and enable the signed manifest. For testing, I'll manually copy
the manifest into some web location and point the redirect to it.

##### Motivation

Part of the upgrader security feature.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
- A future PR (after the client code changes are merged) will leverage
the client code in the pipeline to run in-build tests that ensure that
the client code and the manifest remain in sync.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Run through of all [test
scenarios](https://github.com/Microsoft/accessibility-insights-windows/blob/main/docs/Scenarios.md)
completed?
- [n/a] Does this address an existing issue? If yes, Issue# - 
- [n/a] Includes UI changes?
- [n/a] Run the production version of Accessibility Insights for Windows
against a version with changes.
  - [n/a] Attach any screenshots / GIF's that are applicable.

> Note: After the PR has been created, certain checks will be kicked
off. All of these checks must pass before a merge.